requirements.txt
discord.py>=2.0.0
google-genai>=1.0.0
python-dotenv>=1.0.0
requests>=2.28.0
curl_cffi>=0.7.0
beautifulsoup4>=4.12.0
cloudscraper>=1.2.71
playwright>=1.40.0
line-bot-sdk>=3.0.0
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
networkx>=3.1
# Pixiv 爬取
pixivpy3>=3.7.0
faiss-cpu>=1.7.0
scikit-learn>=1.3.0
ImageHash>=4.3.1
aiohttp>=3.9.0
# 即時狀態頁（Streamlit）+ 固定 ngrok 反向代理
streamlit>=1.30.0
pyngrok>=8.0.0
# SSCD NN binary hash（Phase 3 抗裁剪/修圖/翻譯搜圖）
# 需 Python 3.12；CUDA wheel 另行安裝，不寫在這避免 pip 預設 index：
#   python -m pip install torch>=2.4.0 torchvision>=0.19.0 --index-url https://download.pytorch.org/whl/cu124
torch>=2.4.0
torchvision>=0.19.0
